Move floatToHalf from onnxruntime_utillib to onnxruntime_framework.lib
